requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
jinja2>=3.1.0
google-analytics-data>=0.18.16
google-auth>=2.37.0
google-auth-oauthlib>=1.2.1
google-auth-httplib2>=0.2.0
httpx>=0.28.0
pandas>=2.2.0
numpy>=2.0.0
duckdb>=1.1.0
python-dateutil>=2.9.0
pydantic>=2.10.0
holidays>=0.64
requests>=2.32.0
